1.17.19	MySQLIntegrityConstraintViolationException: Column 'destination_name' cannot be null 1.17.20	Score for bicycling and walking is negative when there is more than 10mm precipitation

+++ sustainify-us-core/src/main/java/us/sustainify/commute/domain/model/score/RouteScore.java
@@ -19,7 +19,7 @@ public abstract class RouteScore {
 
 		@Override
 		protected int forRouteInternal(Route route) {
-			return routeScore.forRoute(route) * (100 / weatherScore.scoreFor(route.getOrigin()));
+			return (int) (routeScore.forRoute(route) * (100d / weatherScore.scoreFor(route.getOrigin())));
 		}
 
 	}
